requirements.txt
asgiref==3.3.0
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-filter==2.4.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.1
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.4
phonenumbers==8.12.12
Pillow==8.0.1
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.11
wrapt==1.12.1
